+++ src-ui/app/components/device-list/device-list.component.ts
@@ -278,7 +278,6 @@ export class DeviceListComponent implements OnInit {
   }
 
   async rightClickBulkPowerLighthouseDevices() {
-    console.log('TRIGGER');
     this.showLHStatePopover = !this.showLHStatePopover;
   }