+++ src-ui/app/services/status-automations/status-change-general-events-automation.service.ts
@@ -70,13 +70,13 @@ export class StatusChangeGeneralEventsAutomationService {
           }
           return { status, statusMessage, sleepMode };
         }),
-        filter((data) => Boolean(data.status || data.statusMessage)),
+        filter((data) => Boolean(data.status !== null || data.statusMessage !== null)),
         debounceTime(500)
       )
       .subscribe(async ({ status, statusMessage, sleepMode }) => {
         const oldStatus = this.vrcUser?.status;
         const oldStatusMessage = this.vrcUser?.statusDescription;
-        const success = await this.vrchat.setStatus(status, statusMessage).catch((e) => false);
+        const success = await this.vrchat.setStatus(status, statusMessage).catch(() => false);
         if (success) {
           if (await this.notifications.notificationTypeEnabled('AUTO_UPDATED_VRC_STATUS')) {
             await this.notifications.send(