+++ src-ui/app/utils/browser-utils.ts
@@ -7,13 +7,4 @@ export function disableDefaultContextMenu() {
     },
     { capture: true }
   );
-
-  document.addEventListener(
-    'selectstart',
-    (e) => {
-      e.preventDefault();
-      return false;
-    },
-    { capture: true }
-  );
 }